refactor(voyage): la persistance et les permaliens sortent d'app.js
`persistance.ts` porte l'encodage de l'état, son écriture dans `localStorage` et dans le hash, sa
relecture, et sa restauration. `app.js` : 3 218 → 3 121 lignes, 20 lignes de globales, et
`restoring` — le verrou anti-resauvegarde — quitte la coquille avec elle.

CE MODULE LIT ET ÉCRIT LE DOM, et c'est la bonne place : 24 des 31 clés sont des CHAMPS, dont la
vérité est le formulaire. Les recopier dans `etat.ts` créerait une seconde vérité sur les réglages
les plus lus — décision prise avec #135, et elle tient ici.

LA DÉCOUPE, encore une fois entre la donnée et l'interface. `applyState` faisait les deux : elle
posait les champs et l'état, PUIS appelait trois synchronisations d'interface (indicateurs de tri,
cases grisées, bouton de board). Ces trois-là restent dans `app.js` et lui sont passées en rappel —
appelé À L'INTÉRIEUR du verrou `restoring`. Ce n'est pas un détail de forme : les laisser à
l'appelant, après le retour, les ferait tourner verrou ouvert, et une seule d'entre elles qui
resauverait écraserait l'état qu'on vient de restaurer.

Contrôle systématique après extraction, désormais réflexe depuis #139 : relever les noms du bloc
encore référencés dans `app.js` et vérifier qu'ils sont TOUS importés. Un identifiant libre ne se
voit ni au build ni au typecheck. Ici : quatre noms, quatre importés. `STATE_KEY` était déclaré des
deux côtés — le doublon est retiré d'`app.js`.

Contribue à #96.
